Update state management to cubit
