+++ doc/confapp_flutter/README.md
@@ -1,6 +1,7 @@
 # confapp_flutter
 
-A new Flutter project.
+- https://github.com/burhanrashid52/WhatTodo/tree/master/lib
+- https://github.com/felangel/bloc/tree/master/examples/flutter_shopping_cart/lib/cart/models
 
 ## Getting Started
