+++ doc/confapp_flutter/lib/app.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+import 'package:confapp_flutter/hcl/hcl.dart';
+
+class HCLApp extends MaterialApp {
+  const HCLApp({super.key}) : super(home: const HCLPage());
+}
