+++ doc/confapp_flutter/lib/hcl/hcl.dart
@@ -1,0 +1,1 @@
+export 'view/hcl_page.dart';
