+++ doc/confapp_flutter/lib/hcl/models/component.dart
@@ -2,14 +2,14 @@ import 'dart:collection';
 
 import 'package:confapp_flutter/components/config.dart';
 import 'package:flutter/foundation.dart';
-import 'package:confapp_flutter/hcl_components.dart';
+import 'package:confapp_flutter/hcl/models/hcl_components.dart';
 
 class ComponentModel extends ChangeNotifier {
   final components = WebPageGenerator();
-  late ConfigGenerator selectedComponent = components.generators[0];
+  late ConfigGenerator selectedComponent = components.components[0];
 
   UnmodifiableListView<ConfigGenerator> get generators =>
-      UnmodifiableListView(components.generators);
+      UnmodifiableListView(components.components);
 
   ConfigGenerator get currComponent => selectedComponent;
