+++ doc/confapp_flutter/lib/hcl/models/hcl_components.dart
@@ -2,7 +2,7 @@ import 'package:confapp_flutter/components/config.dart';
 import 'package:confapp_flutter/components/components.dart';
 
 class WebPageGenerator {
-  final List<ConfigGenerator> generators = [
+  final List<ConfigGenerator> components = [
     RotateGenerator(),
     PriorityArbiterGenerator(),
     RippleCarryAdderGenerator(),
