+++ doc/confapp_flutter/lib/hcl/view/hcl_view.dart.bak
@@ -1,0 +1,45 @@
+import 'package:confapp_flutter/components/config.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:confapp_flutter/hcl/cubit/component_cubit.dart';
+
+class HCLView extends StatelessWidget {
+  const HCLView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final textTheme = Theme.of(context).textTheme;
+    return Scaffold(
+      appBar: AppBar(title: const Text('Counter')),
+      body: Center(
+        child: BlocBuilder<ComponentCubit, ConfigGenerator>(
+          builder: (context, state) {
+            return Text('${state.componentName}',
+                style: textTheme.displayMedium);
+          },
+        ),
+      ),
+      floatingActionButton: Column(
+        mainAxisAlignment: MainAxisAlignment.end,
+        crossAxisAlignment: CrossAxisAlignment.end,
+        children: <Widget>[
+          FloatingActionButton(
+            key: const Key('counterView_increment_floatingActionButton'),
+            child: const Icon(Icons.add),
+            onPressed: () => context
+                .read<ComponentCubit>()
+                .setSelectedComponent(ComponentCubit.generator.components[1]),
+          ),
+          const SizedBox(height: 8),
+          FloatingActionButton(
+            key: const Key('counterView_decrement_floatingActionButton'),
+            child: const Icon(Icons.remove),
+            onPressed: () => context
+                .read<ComponentCubit>()
+                .setSelectedComponent(ComponentCubit.generator.components[2]),
+          ),
+        ],
+      ),
+    );
+  }
+}
