+++ doc/confapp_flutter/pubspec.yaml
@@ -40,6 +40,8 @@ dependencies:
   cupertino_icons: ^1.0.2
   sidebarx: ^0.16.0
   provider: ^6.0.5
+  flutter_bloc: ^8.1.3
+  equatable: ^2.0.5
 
 dev_dependencies:
   flutter_test:
